[Maps] convert HeatmapLayer to TS (#90666)
* [Maps] convert HeatmapLayer to TS

* make VectorLayer implement IVectorLayer

* fix jest test and functional test
