+++ x-pack/plugins/maps/common/descriptor_types/layer_descriptor_types.ts
@@ -8,7 +8,11 @@
 /* eslint-disable @typescript-eslint/consistent-type-definitions */
 
 import { Query } from 'src/plugins/data/public';
-import { StyleDescriptor, VectorStyleDescriptor } from './style_property_descriptor_types';
+import {
+  HeatmapStyleDescriptor,
+  StyleDescriptor,
+  VectorStyleDescriptor,
+} from './style_property_descriptor_types';
 import { DataRequestDescriptor } from './data_request_descriptor_types';
 import { AbstractSourceDescriptor, TermJoinSourceDescriptor } from './source_descriptor_types';
 
@@ -39,4 +43,8 @@ export type LayerDescriptor = {
 
 export type VectorLayerDescriptor = LayerDescriptor & {
   style: VectorStyleDescriptor;
+};
+
+export type HeatmapLayerDescriptor = LayerDescriptor & {
+  style: HeatmapStyleDescriptor;
 };
