+++ x-pack/plugins/maps/public/classes/layers/create_tile_map_layer_descriptor.ts
@@ -23,11 +23,9 @@ import {
   VECTOR_STYLES,
 } from '../../../common/constants';
 import { VectorStyle } from '../styles/vector/vector_style';
-// @ts-ignore
 import { ESGeoGridSource } from '../sources/es_geo_grid_source';
-import { VectorLayer } from './vector_layer/vector_layer';
-// @ts-ignore
-import { HeatmapLayer } from './heatmap_layer/heatmap_layer';
+import { VectorLayer } from './vector_layer';
+import { HeatmapLayer } from './heatmap_layer';
 import { getDefaultDynamicProperties } from '../styles/vector/vector_style_defaults';
 import { NUMERICAL_COLOR_PALETTES } from '../styles/color_palettes';
 import { getSourceAggKey } from '../../../common/get_agg_key';
