+++ x-pack/plugins/maps/public/classes/layers/layer.tsx
@@ -21,6 +21,7 @@ import {
   MAX_ZOOM,
   MB_SOURCE_ID_LAYER_ID_PREFIX_DELIMITER,
   MIN_ZOOM,
+  SOURCE_BOUNDS_DATA_REQUEST_ID,
   SOURCE_DATA_REQUEST_ID,
   SOURCE_TYPES,
   STYLE_TYPE,
@@ -66,6 +67,7 @@ export interface ILayer {
   getImmutableSourceProperties(): Promise<ImmutableSourceProperty[]>;
   renderSourceSettingsEditor({ onChange }: SourceEditorArgs): ReactElement<any> | null;
   isLayerLoading(): boolean;
+  isLoadingBounds(): boolean;
   isFilteredByGlobalTime(): Promise<boolean>;
   hasErrors(): boolean;
   getErrors(): string;
@@ -399,6 +401,11 @@ export class AbstractLayer implements ILayer {
 
   isLayerLoading(): boolean {
     return this._dataRequests.some((dataRequest) => dataRequest.isLoading());
+  }
+
+  isLoadingBounds() {
+    const boundsDataRequest = this.getDataRequest(SOURCE_BOUNDS_DATA_REQUEST_ID);
+    return !!boundsDataRequest && boundsDataRequest.isLoading();
   }
 
   hasErrors(): boolean {
