+++ x-pack/plugins/maps/public/classes/layers/load_layer_wizards.ts
@@ -9,7 +9,6 @@ import { registerLayerWizard } from './layer_wizard_registry';
 import { uploadLayerWizardConfig } from './file_upload_wizard';
 // @ts-ignore
 import { esDocumentsLayerWizardConfig } from '../sources/es_search_source';
-// @ts-ignore
 import { clustersLayerWizardConfig, heatmapLayerWizardConfig } from '../sources/es_geo_grid_source';
 import { geoLineLayerWizardConfig } from '../sources/es_geo_line_source';
 // @ts-ignore
