+++ x-pack/plugins/maps/public/classes/layers/solution_layers/observability/create_layer_descriptor.test.ts
@@ -176,7 +176,6 @@ describe('createLayerDescriptor', () => {
       __dataRequests: [],
       alpha: 0.75,
       id: '12345',
-      joins: [],
       label: '[Performance] Duration',
       maxZoom: 24,
       minZoom: 0,
