+++ x-pack/plugins/maps/public/classes/layers/vector_layer/index.ts
@@ -1,0 +1,9 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export { addGeoJsonMbSource, syncVectorSource } from './utils';
+export { IVectorLayer, VectorLayer, VectorLayerArguments } from './vector_layer';
