+++ x-pack/plugins/maps/public/classes/sources/es_geo_grid_source/clusters_layer_wizard.tsx
@@ -9,10 +9,9 @@ import { i18n } from '@kbn/i18n';
 import React from 'react';
 // @ts-ignore
 import { CreateSourceEditor } from './create_source_editor';
-// @ts-ignore
 import { ESGeoGridSource, clustersTitle } from './es_geo_grid_source';
 import { LayerWizard, RenderWizardArguments } from '../../layers/layer_wizard_registry';
-import { VectorLayer } from '../../layers/vector_layer/vector_layer';
+import { VectorLayer } from '../../layers/vector_layer';
 import {
   ESGeoGridSourceDescriptor,
   ColorDynamicOptions,
