+++ x-pack/plugins/maps/public/classes/sources/es_geo_grid_source/heatmap_layer_wizard.tsx
@@ -9,11 +9,9 @@ import { i18n } from '@kbn/i18n';
 import React from 'react';
 // @ts-ignore
 import { CreateSourceEditor } from './create_source_editor';
-// @ts-ignore
 import { ESGeoGridSource, heatmapTitle } from './es_geo_grid_source';
 import { LayerWizard, RenderWizardArguments } from '../../layers/layer_wizard_registry';
-// @ts-ignore
-import { HeatmapLayer } from '../../layers/heatmap_layer/heatmap_layer';
+import { HeatmapLayer } from '../../layers/heatmap_layer';
 import { ESGeoGridSourceDescriptor } from '../../../../common/descriptor_types';
 import { LAYER_WIZARD_CATEGORY, RENDER_AS } from '../../../../common/constants';
 import { HeatmapLayerIcon } from '../../layers/icons/heatmap_layer_icon';
