+++ x-pack/plugins/maps/public/classes/sources/es_geo_grid_source/index.ts
